Rename spa_ast to ast.
The spa prefix is somewhat redundant.
